Add "Exception" suffix to all exceptions and make them children of CircuitBreakerException
