+++ src/Exceptions/CircuitBreakerException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace PrestaShop\CircuitBreaker\Exceptions;
+
+use Exception;
+
+/**
+ * Base exception for Circuit Breaker exceptions
+ */
+class CircuitBreakerException extends Exception
+{
+
+}
