+++ src/Exceptions/InvalidPlaceException.php
@@ -2,25 +2,22 @@
 
 namespace PrestaShop\CircuitBreaker\Exceptions;
 
-use Exception;
 use PrestaShop\CircuitBreaker\Utils\ErrorFormatter;
 
-final class InvalidTransaction extends Exception
+final class InvalidPlaceException extends CircuitBreakerException
 {
     /**
-     * @param mixed $service the service URI
      * @param mixed $failures the failures
-     * @param mixed $state the Circuit Breaker
+     * @param mixed $timeout the timeout
      * @param mixed $threshold the threshold
      *
      * @return self
      */
-    public static function invalidParameters($service, $failures, $state, $threshold)
+    public static function invalidSettings($failures, $timeout, $threshold)
     {
-        $exceptionMessage = 'Invalid parameters for Transaction' . PHP_EOL .
-            ErrorFormatter::format('service', $service, 'isURI', 'an URI') .
+        $exceptionMessage = 'Invalid settings for Place' . PHP_EOL .
             ErrorFormatter::format('failures', $failures, 'isPositiveInteger', 'a positive integer') .
-            ErrorFormatter::format('state', $state, 'isString', 'a string') .
+            ErrorFormatter::format('timeout', $timeout, 'isPositiveValue', 'a float') .
             ErrorFormatter::format('threshold', $threshold, 'isPositiveInteger', 'a positive integer')
         ;
