--- src/Exceptions/TransactionNotFound.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace PrestaShop\CircuitBreaker\Exceptions;
-
-use Exception;
-
-final class TransactionNotFound extends Exception
-{
-}
