+++ src/Exceptions/TransactionNotFoundException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PrestaShop\CircuitBreaker\Exceptions;
+
+final class TransactionNotFoundException extends CircuitBreakerException
+{
+}
