--- src/Exceptions/UnavailableService.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace PrestaShop\CircuitBreaker\Exceptions;
-
-use Exception;
-
-final class UnavailableService extends Exception
-{
-}
